Changing Block/Item Registration
Changed from Galian's Registry system to my own for better understanding and ease of use.
