+++ src/main/java/org/millenaire/blocks/BlockMillCrops.java
@@ -96,55 +96,4 @@ public class BlockMillCrops extends BlockCrops
 		this.seed = seedIn;
 		return this;
 	}
-	
-    //////////////////////////////////////////////////////////\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
-    
-	//Declarations
-		public static Block cropTurmeric;
-		public static Block cropRice;
-		public static Block cropMaize;
-		public static Block cropGrapeVine;
-		
-		public static Item turmeric;
-		public static Item rice;
-		public static Item maize;
-		public static Item grapes;
-
-    public static void preinitialize()
-    {
-    	cropTurmeric = new BlockMillCrops(false, false).setCreativeTab(null).setUnlocalizedName("cropTurmeric");
-    	turmeric = new ItemMillSeeds(cropTurmeric).setCreativeTab(Millenaire.tabMillenaire).setUnlocalizedName("turmeric");
-    	((BlockMillCrops) cropTurmeric).setSeed((IPlantable) turmeric);
-    	GameRegistry.registerItem(turmeric, "turmeric");
-    	GameRegistry.registerBlock(cropTurmeric, "cropTurmeric");
-    	
-    	cropRice = new BlockMillCrops(true, false).setCreativeTab(null).setUnlocalizedName("cropRice");
-    	rice = new ItemMillSeeds(cropRice).setCreativeTab(Millenaire.tabMillenaire).setUnlocalizedName("rice");
-    	((BlockMillCrops) cropRice).setSeed((IPlantable) rice);
-    	GameRegistry.registerItem(rice, "rice");
-    	GameRegistry.registerBlock(cropRice, "cropRice");
-
-    	cropMaize = new BlockMillCrops(false, true).setCreativeTab(null).setUnlocalizedName("cropMaize");
-    	maize = new ItemMillSeeds(cropMaize).setCreativeTab(Millenaire.tabMillenaire).setUnlocalizedName("maize");
-    	((BlockMillCrops) cropMaize).setSeed((IPlantable) maize);
-    	GameRegistry.registerItem(maize, "maize");
-    	GameRegistry.registerBlock(cropMaize, "cropMaize");
-
-    	cropGrapeVine = new BlockMillCrops(false, false).setCreativeTab(null).setUnlocalizedName("cropGrapeVine");
-    	grapes = new ItemMillSeeds(cropGrapeVine).setCreativeTab(Millenaire.tabMillenaire).setUnlocalizedName("grapes");
-    	((BlockMillCrops) cropGrapeVine).setSeed((IPlantable) grapes);
-    	GameRegistry.registerItem(grapes, "grapes");
-    	GameRegistry.registerBlock(cropGrapeVine, "cropGrapeVine");
-    }
-    
-    @SideOnly(Side.CLIENT)
-	public static void render()
-	{
-		RenderItem renderItem = Minecraft.getMinecraft().getRenderItem();
-		
-		renderItem.getItemModelMesher().register(turmeric, 0, new ModelResourceLocation(Millenaire.MODID + ":turmeric", "inventory"));
-		renderItem.getItemModelMesher().register(rice, 0, new ModelResourceLocation(Millenaire.MODID + ":rice", "inventory"));
-		renderItem.getItemModelMesher().register(maize, 0, new ModelResourceLocation(Millenaire.MODID + ":maize", "inventory"));
-		renderItem.getItemModelMesher().register(grapes, 0, new ModelResourceLocation(Millenaire.MODID + ":grapes", "inventory"));
-	}
 }
