+++ src/main/java/org/millenaire/items/ItemMillSign.java
@@ -2,23 +2,17 @@ package org.millenaire.items;
 
 import org.millenaire.Millenaire;
 import org.millenaire.blocks.BlockMillSign;
+import org.millenaire.blocks.MillBlocks;
 
-import net.minecraft.block.BlockStandingSign;
-import net.minecraft.block.BlockWallSign;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.entity.RenderItem;
 import net.minecraft.client.resources.model.ModelResourceLocation;
-import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.init.Blocks;
 import net.minecraft.item.Item;
-import net.minecraft.item.ItemBlock;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.tileentity.TileEntitySign;
 import net.minecraft.util.BlockPos;
 import net.minecraft.util.EnumFacing;
-import net.minecraft.util.MathHelper;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.common.registry.GameRegistry;
 import net.minecraftforge.fml.relauncher.Side;
@@ -58,7 +52,7 @@ public class ItemMillSign extends Item
 			}
 			else
 			{
-				worldIn.setBlockState(pos, BlockMillSign.blockMillSign.getDefaultState().withProperty(BlockMillSign.FACING, side)/*Blocks.wall_sign.getDefaultState().withProperty(BlockWallSign.FACING, side)*/, 3);
+				worldIn.setBlockState(pos, MillBlocks.blockMillSign.getDefaultState().withProperty(BlockMillSign.FACING, side)/*Blocks.wall_sign.getDefaultState().withProperty(BlockWallSign.FACING, side)*/, 3);
 
 
 				--stack.stackSize;
